+++ src/ai/agents/tutor-agent.js
@@ -1,5 +1,5 @@
 import { ASK_USER_TOOL_NAME } from '../tools/ask-user.js';
-import Agent from '../agent.js';
+import Agent from './agent.js';
 import { WORDPRESS_TUTOR_AGENT_ID } from './default-agents.js';
 import { DotPromptTemplate } from '../prompt-template.js';
